refactor(parsers): Move L7 domain logic from ramses_tx to ramses_rf (#751)
